split state stuff; added wifi_loop
