+++ include/state_handler.h
@@ -1,0 +1,9 @@
+#ifndef STATE_HANDLER_H
+#define STATE_HANDLER_H
+#include <WiFi.h>
+#include "mqtt_handler.h"
+#include <map>
+
+void state_handler();
+
+#endif //STATE_HANDLER_H
