+++ include/version.h
@@ -1,1 +1,1 @@
-#define VERSION "0.1.9"
+#define VERSION "0.1.10"
