+++ include/wifi_handler.h
@@ -4,8 +4,10 @@
 #include <Arduino.h>
 #include <WiFi.h>
 #include "macros.h"
+#include "state_handler.h"
 
 void init_wifi();
+void wifi_loop();
 
 // DHCP Hostname
 #define WIFI_DHCPNAME TEXTIFY(CLTNAME)
