+++ src/mqtt_handler.cpp
@@ -19,48 +19,9 @@ String willMessage = "offline";
 byte willQoS = 0;
 boolean willRetain = true;
 
-// Define the map to store key-value pairs
-std::map<String, String> stateMap;
-
-String formatUptime(time_t uptime) {
-    int days = uptime / 86400;
-    int hours = (uptime % 86400) / 3600;
-    int minutes = (uptime % 3600) / 60;
-    int secs = uptime % 60;
-
-    char buffer[20];
-    snprintf(buffer, sizeof(buffer), "%dd %02d:%02d:%02d", days, hours, minutes, secs);
-    return String(buffer);
-}
-
-void setState(String key, String value, bool publish) {
-    stateMap[key] = value;
-    String fullTopic = mqttname + "/status/" + key;
-    if (publish && mqtt_client.connected()) {
-        mqtt_client.publish(fullTopic.c_str(), value.c_str());
-    }
-}
-
-void publishStates() {
-    for (const auto &kv : stateMap) {
-        String fullTopic = mqttname + "/status/" + kv.first;
-        mqtt_client.publish(fullTopic.c_str(), kv.second.c_str());
-    }
-}
-
-void publishStatesTask(void *pvParameters) {
-    while (true) {
-
-        time_t now = time(nullptr);
-        time_t bootTime = now - esp_timer_get_time() / 1000000;
-        setState("uptime", formatUptime(now - bootTime), false);
-
-        if (mqtt_client.connected()) {
-            publishStates();
-        }
-        vTaskDelay(min_publish_time * 1000 / portTICK_PERIOD_MS); // Delay for min_publish_time seconds
-    }
-}
+WiFiClient wifi_client;
+void MQTTCallback(char *topic, byte *payload, unsigned int length);
+PubSubClient mqtt_client(mqtt_server, mqtt_port, MQTTCallback, wifi_client);
 
 void handleMQTTSettingsMessage(const char *topic, const char *command, byte *payload, unsigned int length, bool &flag) {
     // Find the last occurrence of '/'
@@ -120,9 +81,6 @@ void MQTTCallback(char *topic, byte *payload, unsigned int length) {
     handleMQTTMessage(topic, String("min_publish_time").c_str(), payload, length, debug_flg_full_log, true);
 }
 
-WiFiClient wifi_client;
-PubSubClient mqtt_client(mqtt_server, mqtt_port, MQTTCallback, wifi_client);
-
 // Reconnect to MQTT broker
 boolean reconnect() {
     mqtt_buffer_maxed = mqtt_client.setBufferSize(MQTT_BUFFER_SIZE);
@@ -158,7 +116,6 @@ boolean reconnect() {
         set_led(LedState::LED_FLASH);
 #endif
     }
-
     return mqtt_client.connected();
 }
 
@@ -171,6 +128,7 @@ void mqtt_loop() {
 
             lastReconnectAttempt = now;
             if (reconnect()) {
+                mqtt_client.loop();
                 lastReconnectAttempt = 0;
                 DEBUG_PRINTLN("MQTT Reconnected.");
             } else {
@@ -179,7 +137,6 @@ void mqtt_loop() {
             }
         }
     } else {
-
         mqtt_client.loop();
     }
 }
@@ -188,14 +145,7 @@ void mqtt_init() {
     DEBUG_PRINTLN("Connecting MQTT ...");
     if (reconnect()) {
         DEBUG_PRINTLN("MQTT Connected.");
-        // Create the task to call publishStates() every min_publish_time seconds
-        xTaskCreate(publishStatesTask, "Publish States Task", 2048, NULL, 0, NULL);
-
-        setState("version", VERSION, true);
-        setState("ipaddress", WiFi.localIP().toString(), true);
-        setState("ble_connection", "startup", true);
-        setState("status", "online", true);
-
-    } else
+    } else {
         DEBUG_PRINTLN("MQTT Connect failed.");
+    }
 }
