+++ src/wifi_handler.cpp
@@ -28,3 +28,11 @@ void init_wifi() {
     DEBUG_PRINT("IP address: ");
     DEBUG_PRINTLN(WiFi.localIP());
 }
+
+void wifi_loop() {
+    if (WiFi.status() != WL_CONNECTED) {
+        DEBUG_PRINTLN("WiFi not connected, attempting to reconnect...");
+        init_wifi();
+        setState("ipaddress", WiFi.localIP().toString(), false);
+    }
+}
